Scraping
Scraping mostly finished and ready to test, need to work over analysis
side of things.
