+++ RB.OLD/CollocateLocator.py
