+++ RB.OLD/insultcounter.py
