+++ RB.OLD/sentimentAverage.py
