--- badwords.txt
@@ -1,494 +1,0 @@
-ahole, asshole
-anus
-ash0le, asshole
-ash0les, asshole
-asholes, asshole
-ass
-Ass Monkey, asshole
-Assface, asshole
-assh0le, asshole
-assh0lez, asshole
-asshole
-assholes
-assholz, asshole
-asswipe, asshole
-azzhole, asshole
-bassterds, basterd
-bastard
-bastards
-bastardz, basterd
-basterds, basterd
-basterdz, basterd
-Biatch, bitch
-bitch
-bitches, bitch
-Blow Job, blowjob
-boffing
-butthole, asshole
-buttwipe, asshole
-c0ck, cock
-c0cks, cock
-c0k, cock
-Carpet Muncher
-cawk, cock
-cawks, cock
-Clit
-cnts, cunt
-cntz, cunt
-cock
-cockhead, cock
-cock-head, cock
-cocks
-CockSucker
-cock-sucker, cocksucker
-crap
-cum
-cunt
-cunts
-cuntz, cunt
-dick
-dild0, dildo
-dild0s, dildo
-dildo
-dildos
-dilld0, dildo
-dilld0s, dildo
-dominatricks, dominatrix
-dominatrics, dominatrix
-dominatrix
-dyke
-enema
-f u c k, fuck
-f u c k e r, fucker
-fag
-fag1t, faggot
-faget, faggot
-fagg1t, faggot
-faggit, faggot
-faggot, faggot
-fagit, faggot
-fagget, faggot
-fags
-fagz, fag
-faig, fag
-faigs, fag
-fart
-flipping the bird
-fuck
-fucker
-fuckin, fucking
-fucking
-fucks
-Fudge Packer
-fuk, fuck
-Fukah, fucker
-Fuken, fucking
-fuker, fucker
-Fukin, fucking
-Fukk, fuck
-Fukkah, fucker
-Fukken, fucking
-Fukker, fucker
-Fukkin, fucking
-g00k, gook
-gay
-gayboy, gay
-gaygirl, gay
-gays
-gayz, gay
-God-damned, goddamn
-h00r, whore
-h0ar, whore
-h0re, whore
-hells
-hoar, whore
-hoorm whore
-hoorem whore
-jackoff
-jap
-japs
-jerk-off
-jisim
-jiss
-jizm
-jizz
-knob
-knobs
-knobz, knob
-kunt, cunt
-kunts, cunt
-kuntz, cunt
-Lesbian
-Lezzian, lesbian
-Lipshits 
-Lipshitz, lipshits
-masochist
-masokist, masochist
-massterbait, masturbate
-masstrbait, masturbate
-masstrbate, masturbate
-masterbaiter, masturbate
-masterbate, masturbate
-masterbates, masturbate
-Motha Fucker, motherfucker 
-Motha Fuker, motherfucker
-Motha Fukkah, motherfucker
-Motha Fukker, motherfucker
-Mother Fucker, motherfucker
-Mother Fukah, motherfucker
-Mother Fuker, motherfucker
-Mother Fukkah, motherfucker
-Mother Fukker, motherfucker
-mother-fucker, motherfucker
-Mutha Fucker, motherfucker
-Mutha Fukah, motherfucker
-Mutha Fuker, motherfucker
-Mutha Fukkah, motherfucker
-Mutha Fukker, motherfucker
-n1gr, nigger
-nastt
-nigger
-nigur, nigger
-niiger, nigger
-niigr, nigger
-orafis, orifice
-orgasim, orgasm
-orgasm
-orgasum, orgasm
-oriface, orifice
-orifice
-orifiss, orifice
-packi, paki
-packie, paki
-packy, paki
-paki
-pakie, paki
-paky, paki
-pecker
-peeenus, penis
-peeenusss, penis
-peenus, penis
-peinus, penis
-pen1s, penis
-penas, penis
-penis
-penis-breath, penis
-penus, penis
-penuus, penis
-Phuc, fuck
-Phuck, fuck
-Phuk, fuck
-Phuker, fuck
-Phukker, fucker
-polac, polak
-polack, polak
-polak
-Poonani
-pr1c, prick
-pr1ck, prick
-pr1k, prick
-pusse, pussy
-pussee, pussy
-pussy
-puuke, puke
-puuker, puker
-queer
-queers
-queerz, queer
-qweers, queer
-qweerz, queer
-qweir, queer
-recktum, rectum
-rectum
-retard
-sadist
-scank, skank
-schlong
-screwing
-semen
-sex
-sexy
-Sh!t, shit
-sh1t, shit
-sh1ter, shit
-sh1ts, shit
-sh1tter, shit
-sh1tz, shit
-shit
-shits
-shitter, shit
-Shitty, shit
-Shity, shit
-shitz, shit
-Shyt, shit
-Shyte, shit
-Shytty, shit
-Shyty, shit
-skanck, skank
-skank
-skankee, skank
-skankey, skank
-skanks, skank
-Skanky, skank
-slut
-sluts, slut
-Slutty, slut
-slutz, slut
-son-of-a-bitch, sob
-tit
-turd
-va1jina, vagina
-vag1na, vagina
-vagiina, vagina
-vagina
-vaj1na, vagina
-vajina, vagina
-vullva, vulva
-vulva
-w0p, wop
-wh00r. whore
-wh0re, whore
-whore
-xrated
-xxx
-b!+ch, bitch
-bitch
-blowjob
-clit
-arschloch
-fuck
-shit
-ass
-asshole
-b!tch, bitch
-b17ch, bitch
-b1tch, bitch
-bastard
-bi+ch, bitch
-boiolas
-buceta
-c0ck, cock
-cawk, cock
-chink
-cipa
-clits
-cock
-cum
-cunt
-dildo
-dirsa
-ejakulate, ejaculate
-fatass, fat ass
-fcuk, fuck
-fuk, fuck
-fux0r, fucker
-hoer, whore
-hore, whore
-jism
-kawk, cock
-l3itch
-l3i+ch
-lesbian
-masturbate
-masterbat*, masturbate
-masterbat3, masturbate
-motherfucker
-s.o.b., sob
-mofo, motherfucker
-nazi
-nigga, nigger
-nigger
-nutsack
-phuck, fuck
-pimpis
-pusse, pussy
-pussy
-scrotum
-sh!t, shit
-shemale
-shi+, shit
-sh!+, shit
-slut
-smut
-teets
-tits
-boobs
-b00bs, boobs
-teez
-testical
-testicle
-titt, tit
-w00se, wuss
-jackoff
-wank
-whoar, whore
-whore
-*damn
-*dyke
-*fuck*
-*shit*
-@$$, ass
-amcik
-andskota
-arse*
-assrammer
-ayir
-bi7ch, bitch
-bitch*
-bollock*
-breasts
-butt-pirate
-cabron
-cazzo
-chraa
-chuj
-Cock*
-cunt*
-d4mn, damn
-daygo
-dego
-dick*
-dike*
-dupa
-dumbfuck, fuck
-dumbass, ass
-dziwka
-ejackulate, ejaculate
-Ekrem*
-Ekto
-enculer
-faen
-fag*
-fanculo
-fanny
-feces
-feg
-Felcher
-ficken, fucking
-fitt*
-Flikker
-foreskin
-Fotze
-Fu(*, fuck
-fuk*
-futkretzn
-gay
-gook
-guiena
-h0r, whore
-h4x0r, hacker
-hell
-helvete
-hoer*
-honkey
-Huevon
-hui
-injun
-jizz
-kanker*
-kike
-klootzak
-kraut
-knulle
-kuk
-kuksuger
-Kurac
-kurwa
-kusi*
-kyrpa*
-lesbo, lesbian
-mamhoon
-masturbat*
-merd*
-mibun
-monkleigh
-mouliewop
-muie
-mulkku
-muschi
-nazis
-nepesaurio
-nigger*
-orospu
-paska*
-perse
-picka
-pierdol*
-pillu*
-pimmel
-piss*
-pizda
-poontsee
-poop
-porn
-p0rn, porn
-pr0n, porn
-preteen
-pula
-pule
-puta
-puto
-qahbeh
-queef*
-rautenberg
-schaffer
-scheiss*
-schlampe
-schmuck
-screw
-sh!t*, shit
-sharmuta
-sharmute
-shipal
-shiz
-skribz
-skurwysyn
-sphencter
-spic
-spierdalaj
-splooge
-suka
-b00b*
-testicle*
-titt*, tit
-twat
-vittu
-wank*
-wetback*
-wichser
-wop*
-yed
-zabourah
-d*ck, dick
-d!ck, dick
-dik, dick
-b*tch, bitch
-b!tch, bitch
-f.ck, fuck
-fu, fuck you
-f you, fuck you
-f uck, fuck
-fu ck, fuck
-fuc, fuck
-fcuk, fuck
-fck, fuck
-fuk, fuck
-f ing, fuck
-fuckin, fuck
-sh#t, shit
-sh1t, shit
-sh!t, shit
-sux, sucks
-azz, ass
-@ss, ass
-8ss, ass
-a$$, ass
-a s s, ass
-ass-hole, asshole
-a-hole, asshole
-w t f, wtf
-wtf, what the fuck
-s t f u, stfu
-s.o.b, sob
-idi0t, idiot
-nigga, nigger
