+++ resbot.ini
@@ -1,0 +1,19 @@
+[IMPORTANT]
+DATA_DIR: C:/test/directory/
+LAST_UPDATED:
+
+[PRAW]
+
+[REDDIT]
+SFW: funny,aww,gaming
+NSFW:
+
+[4CHAN]
+SFW: /g/,/v/,/vg/
+NSFW: /b/
+
+[CORRELATIONS]
+VIDEOGAMES: /g/,/vg/,gaming
+
+[CORPORA]
+MASTER_CORPUS: C:/test/directory/master_corpus.csv
